Split `type_check.rs` into smaller modules.
